This contains a lot of documentation edits to make docstrings sphinx-compatible.
Not as much has been done with cdutil as was done in genutil, but much of the
documentation has been converted to RST format, there are :Example: sections
in a lot of places, and there is a sphinx configuration that makes it possible
to build pdf/html output for the documentation.

As with genutil, cdutil needs an experienced eye to go over the documentation
and fill in parameter/return descriptions/datatypes. It also needs working
:Examples: that will pass doctests.